RuntimeStatus is a metric for saying something about the runtime-state of the application.
You might have an application that starts up fairly quicky but then
has to fetch loads of data before it is actually ready to receive load.
For instance: With this metric you can see and check that your loadbalancer
is doing the correct thing and only sending requests when online
or you can correlate state with api-errors.