Fix a classic gotcha when using Micrometer where the value you pass in gets GC-ed in runtime
Micrometer gauges does not retain the actual object you pass in for it
 self. So as long as that object is not GC-ed, you are good.
In this implementation, I thought that I was in the clear since
the `OneMinuteEventsThreshold.perMinute(1)` usage is in an enum, which
is instantiated as singletons and thus is not GC-ed.

The gotcha is of course that when we return the primitive `double`, is
is not a reference but a copy of the actual bit that represent the primitive.
Thus, this gets GC-ed.

Anyhooo, By calling `Runtime.getRuntime().gc();` in the test we can
verify this behavour.