Fix bug where exception were swallowed if AppStatus was not FAILED.
Exception should be thrown without regards to the appStatus. The old flow suggested that an exception only was expected to be thrown if the predicate judged the call to be AppStatus.Failed (based on returnValue and the exception). This may not always be the case e.g. an network exception where the exception may act as a retry-flag and we don't want to alert on it (i.e. AppStatus.OK)