Allow creating AppBusinessEvent with lambda
If you're not interested in alarm thresholds, it is nice to be able to
create AppBusinessEvents with lambda expressions, which will only
implement the getName() method. This is not possible when you have to
implement all three methods. Since the thresholds return Optionals, I
think it makes sense to make default implementations that return
Optional.empty().